backend/requirements.txt
fastapi
uvicorn
cohere
supabase
python-dotenv
pydantic
langgraph
langgraph-checkpoint-sqlite
aiosqlite
pymupdf
pytesseract
pillow
python-docx
python-multipart
pytest
pytest-asyncio
httpx
pytest-cov
